Extract animation state to custom hook
